# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger api for Kaho BaaS",
        "title": "Kaho BaaS API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/account": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get Detail Account",
                "responses": {}
            }
        },
        "/account/sessions/current": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Logout",
                "responses": {}
            }
        },
        "/account/sessions/email": {
            "post": {
                "description": "Untuk email, bisa menggunakan email atau username",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email or username",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}